pyproject.toml
[project]
name = "athlete-dash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "plotly>=6.6.0",
    "polars>=1.39.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.6",
    "stravalib>=2.4",
    "streamlit>=1.55.0",
]
